+++ fake-server/package.json
@@ -1,4 +1,9 @@
 {
+  "scripts": {
+    "start": "ts-node server.ts",
+    "test": "mocha",
+    "test:watch": "mocha --watch"
+  },
   "dependencies": {
     "@types/body-parser": "^1.19.5",
     "@types/express": "^4.17.21",
@@ -7,5 +12,11 @@
     "express": "^4.18.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.3.2"
+  },
+  "devDependencies": {
+    "@types/mocha": "^10.0.6",
+    "@types/supertest": "^2.0.16",
+    "mocha": "^10.2.0",
+    "supertest": "^6.3.3"
   }
 }
